Shift Duke app under package
Duke app not under a package which does not follow requirements.

Duke app now placed under Duke folder.
